Add black line center detection code to linefollower.cpp

+++ src/GUI/include/filters.h
@@ -49,7 +49,7 @@ public:
 class BlackLineCenter : public Filter {
 private:
 	static std::string name;
-	int thVal;
+	double thVal;
 	double areaThresh;
 public:
 	BlackLineCenter();
